move getInternshipFromTrackedId to api.js
