+++ client/src/utils/helper.jsx
@@ -4,8 +4,6 @@ import SchoolOutlinedIcon from '@mui/icons-material/SchoolOutlined';
 import WorkOutlineRoundedIcon from '@mui/icons-material/WorkOutlineRounded';
 import React from 'react';
 
-import { MOCK_INTERNSHIP_DATA } from './mockData';
-
 // returns the correct icon depending on the corresponding filter the tag is showing
 export const getLabelIcon = (filter) => {
   switch (filter) {
@@ -19,9 +17,3 @@ export const getLabelIcon = (filter) => {
       return <CheckCircleOutlineRoundedIcon />;
   }
 };
-
-// TODO: replace with database query once backend is completed
-export const getInternshipFromTrackedId = (id) =>
-  MOCK_INTERNSHIP_DATA.find((mockInternship) => {
-    return mockInternship.id === id;
-  });
